+++ src/main/java/mc/replay/wrapper/entity/metadata/MobMetadata.java
@@ -7,8 +7,8 @@ import org.jetbrains.annotations.Nullable;
 
 public class MobMetadata extends LivingEntityMetadata {
 
-    public static final byte OFFSET = LivingEntityMetadata.MAX_OFFSET;
-    public static final byte MAX_OFFSET = OFFSET + 1;
+    public static final int OFFSET = LivingEntityMetadata.MAX_OFFSET;
+    public static final int MAX_OFFSET = OFFSET + 1;
 
     private final static byte NO_AI_BIT = 0x01;
     private final static byte IS_LEFT_HANDED_BIT = 0x02;
