+++ src/main/java/mc/replay/wrapper/entity/metadata/animal/CowMetadata.java
@@ -7,8 +7,8 @@ import org.jetbrains.annotations.Nullable;
 
 public class CowMetadata extends AnimalMetadata {
 
-    public static final byte OFFSET = AnimalMetadata.MAX_OFFSET;
-    public static final byte MAX_OFFSET = OFFSET + 0;
+    public static final int OFFSET = AnimalMetadata.MAX_OFFSET;
+    public static final int MAX_OFFSET = OFFSET + 0;
 
     public CowMetadata(@Nullable EntityWrapper entity, @NotNull Metadata metadata) {
         super(entity, metadata);
