+++ src/main/java/mc/replay/wrapper/entity/metadata/animal/PigMetadata.java
@@ -7,8 +7,8 @@ import org.jetbrains.annotations.Nullable;
 
 public class PigMetadata extends AnimalMetadata {
 
-    public static final byte OFFSET = AnimalMetadata.MAX_OFFSET;
-    public static final byte MAX_OFFSET = OFFSET + 2;
+    public static final int OFFSET = AnimalMetadata.MAX_OFFSET;
+    public static final int MAX_OFFSET = OFFSET + 2;
 
     public PigMetadata(@Nullable EntityWrapper entity, @NotNull Metadata metadata) {
         super(entity, metadata);
