+++ src/main/java/mc/replay/wrapper/entity/metadata/animal/tameable/ParrotMetadata.java
@@ -7,8 +7,8 @@ import org.jetbrains.annotations.Nullable;
 
 public class ParrotMetadata extends TameableAnimalMetadata {
 
-    public static final byte OFFSET = TameableAnimalMetadata.MAX_OFFSET;
-    public static final byte MAX_OFFSET = OFFSET + 1;
+    public static final int OFFSET = TameableAnimalMetadata.MAX_OFFSET;
+    public static final int MAX_OFFSET = OFFSET + 1;
 
     public ParrotMetadata(@Nullable EntityWrapper entity, @NotNull Metadata metadata) {
         super(entity, metadata);
