+++ src/main/java/mc/replay/wrapper/entity/metadata/item/FireballMetadata.java
@@ -10,8 +10,8 @@ import org.jetbrains.annotations.Nullable;
 
 public class FireballMetadata extends ItemContainingMetadata implements ObjectDataProvider, ProjectileMetadata {
 
-    public static final byte OFFSET = ItemContainingMetadata.MAX_OFFSET;
-    public static final byte MAX_OFFSET = OFFSET + 0;
+    public static final int OFFSET = ItemContainingMetadata.MAX_OFFSET;
+    public static final int MAX_OFFSET = OFFSET + 0;
 
     private EntityWrapper shooter;
