+++ src/main/java/mc/replay/wrapper/entity/metadata/item/ItemContainingMetadata.java
@@ -11,8 +11,8 @@ import org.jetbrains.annotations.Nullable;
 
 class ItemContainingMetadata extends EntityMetadata {
 
-    public static final byte OFFSET = EntityMetadata.MAX_OFFSET;
-    public static final byte MAX_OFFSET = OFFSET + 1;
+    public static final int OFFSET = EntityMetadata.MAX_OFFSET;
+    public static final int MAX_OFFSET = OFFSET + 1;
 
     private final ItemWrapper defaultItem;
